Serialize everything.
Doing it in parallel in the natural node-ish way makes LJ complain that
the client is making repeated requests. E.g., the userpic fetcher might
be requesting a challenge at about the same time the event fetcher is,
because one is running while the other is waiting on network events.